% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_functions.R
\name{bbsummary}
\alias{bbsummary}
\title{bbsummary}
\usage{
bbsummary(M)
}
\arguments{
\item{M}{object of class "bbreg" containing results from the fitted model.}
}
\description{
Function providing a summary of results related to the regression model (bessel or beta).
}
\examples{
\donttest{
fit1 = bbreg(agreement ~ priming + eliciting, data = WT)
bbsummary(fit1)}
}
\seealso{
\code{\link{eplot}}, \code{\link{dbbtest}}
}
