% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_Bes.R
\name{bessel}
\alias{bessel}
\alias{dbessel}
\alias{rbessel}
\alias{pbessel}
\alias{qbessel}
\title{Bessel Distribution}
\usage{
dbessel(z, mu = 1/2, phi = 1)

rbessel(n, mu = 1/2, phi = 1)

pbessel(q, mu = 1/2, phi = 1)

qbessel(p, mu = 1/2, phi = 1)
}
\arguments{
\item{z}{vector of numbers in (0,1) for which the p.d.f. is to be evaluated.}

\item{mu}{vector of numbers in (0,1) representing the mean parameter.}

\item{phi}{vector of positive numbers representing the precision parameter.}

\item{n}{sample size.}

\item{q}{vector of quantiles.}

\item{p}{vector of probabilities.}
}
\value{
scalar expressing the value of the density at z.
}
\description{
Function to calculate the cumulative distribution, probability density,
generate bessel random numbers and find quantiles of the bessel distribution
}
\examples{
z <- seq(0.01, 0.99, 0.01)
np <- length(z)
density <- rep(0, np)
for (i in 1:np) {
  density[i] <- dbessel(z[i], 0.5, 1)
}
plot(z, density, type = "l", lwd = 2, cex.lab = 2, cex.axis = 2)
}
\seealso{
\code{\link{simdata_bes}}, \code{\link{dbbtest}}, \code{\link{simdata_bet}}
}
