% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_functions.R
\name{d2mudeta2}
\alias{d2mudeta2}
\title{d2mudeta2}
\usage{
d2mudeta2(link.mean, mu)
}
\arguments{
\item{link.mean}{a string containing the link function for the mean.
The possible link functions for the mean are "logit","probit", "cauchit", "cloglog".}

\item{mu}{mean parameter.}
}
\description{
Function to obtain the second derivatives of the mean parameter with respect to the linear predictor.
}
