% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_functions.R
\name{d2phideta2}
\alias{d2phideta2}
\title{d2phideta2}
\usage{
d2phideta2(link.precision, phi)
}
\arguments{
\item{link.precision}{a string containing the link function the precision parameter.
The possible link functions for the precision parameter are "identity", "log", "sqrt", "inverse".}

\item{phi}{precision parameter.}
}
\description{
Function to obtain the second derivatives of the precision parameter with respect to the linear predictor.
}
