% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_Bes.R
\name{D2Q_Obs_Fisher_bes}
\alias{D2Q_Obs_Fisher_bes}
\title{D2Q_Obs_Fisher_bes}
\usage{
D2Q_Obs_Fisher_bes(theta, z, x, v, link.mean, link.precision)
}
\arguments{
\item{theta}{vector of parameters (all coefficients: kappa and lambda).}

\item{z}{response vector with 0 < z_i < 1.}

\item{x}{matrix containing the covariates for the mean submodel. Each column is a different covariate.}

\item{v}{matrix containing the covariates for the precision submodel. Each column is a different covariate.}

\item{link.mean}{a string containing the link function for the mean.
The possible link functions for the mean are "logit","probit", "cauchit", "cloglog".}

\item{link.precision}{a string containing the link function the precision parameter.
The possible link functions for the precision parameter are "identity", "log", "sqrt", "inverse".}
}
\value{
Hessian of the Q-function.
}
\description{
Auxiliary function to compute the observed Fisher information matrix for the bessel regression.
}
