% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_em_dbb.R
\name{dbbtest}
\alias{dbbtest}
\title{dbbtest}
\usage{
dbbtest(
  formula,
  data,
  link.mean,
  link.precision,
  em_controls = list(maxit = 5000, em_tol = 10^(-5)),
  optim_method = "L-BFGS-B",
  optim_controls = list()
)
}
\arguments{
\item{formula}{symbolic description of the model (set: z ~ x or z ~ x | v); see details below.}

\item{data}{arguments considered in the formula description. This is usually a data frame composed by:
(i) the response with bounded continuous observations (0 < z_i < 1),
(ii) covariates for the mean submodel (columns of matrix x) and
(iii) covariates for the precision submodel (columns of matrix v).}

\item{link.mean}{a string containing the link function for the mean.
The possible link functions for the mean are "logit","probit", "cauchit", "cloglog".}

\item{link.precision}{a string containing the link function the precision parameter.
The possible link functions for the precision parameter are "identity", "log", "sqrt".}

\item{em_controls}{a list containing two elements: \code{maxit} that contains the maximum number of iterations of the EM algorithm, the default is set to 5000;
\code{em_tol} that defines the tolerance value to control the convergence criterion in the EM-algorithm, the default is set to 10^(-5).}

\item{optim_method}{main optimization algorithm to be used. The available methods are the same as those of \code{optim} function. The default is set to "L-BFGS-B".}

\item{optim_controls}{a list of control arguments to be passed to the \code{optim} function in the optimization of the model. For the control options, see
the 'Details' in the help of \code{\link[stats]{optim}} for the possible arguments.}
}
\value{
Object of class dbbtest, which is a list containing two elements. The 1st one is a table of terms
considered in the decision rule of the test; they are sum(z2/n) = sum_{i=1}^{n}(z_i^2)/n, sum(quasi_mu) = sum_{i=1}^{n}(tilde{mu_i}^2 + tilde{mu_i}(1-tilde{mu_i})/2)
|D_bessel| and |D_beta| as indicated in the main reference. The 2nd term of the list is the name of the selected model (bessel or beta).
}
\description{
Function to run the discrimination test between beta and bessel regressions (DBB).
}
\examples{
# Illustration using the Weather task data set available in the bbreg package.
dbbtest(agreement ~ priming + eliciting,
  data = WT,
  link.mean = "logit", link.precision = "identity"
)
}
\seealso{
\code{\link{simdata_bes}}, \code{\link{dbessel}}, \code{\link{simdata_bet}}
}
