% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_em_beta.R
\name{EMrun_bet}
\alias{EMrun_bet}
\title{EMrun_bet}
\usage{
EMrun_bet(
  kap,
  lam,
  z,
  x,
  v,
  link.mean,
  link.precision,
  em_controls = list(maxit = 5000, em_tol = 10^(-5)),
  optim_method = "L-BFGS-B",
  optim_controls = list()
)
}
\arguments{
\item{kap}{initial values for the coefficients in kappa related to the mean parameter.}

\item{lam}{initial values for the coefficients in lambda related to the precision parameter.}

\item{z}{response vector with 0 < z_i < 1.}

\item{x}{matrix containing the covariates for the mean submodel. Each column is a different covariate.}

\item{v}{matrix containing the covariates for the precision submodel. Each column is a different covariate.}

\item{link.mean}{a string containing the link function for the mean.
The possible link functions for the mean are "logit","probit", "cauchit", "cloglog".}

\item{link.precision}{a string containing the link function the precision parameter.
The possible link functions for the precision parameter are "identity", "log", "sqrt".}

\item{em_controls}{a list containing two elements: \code{maxit} that contains the maximum number of iterations of the EM algorithm, the default is set to 5000;
\code{em_tol} that defines the tolerance value to control the convergence criterion in the EM-algorithm, the default is set to 10^(-5).}

\item{optim_method}{main optimization algorithm to be used. The available methods are the same as those of \code{optim} function. The default is set to "L-BFGS-B".}

\item{optim_controls}{a list of control arguments to be passed to the \code{optim} function in the optimization of the model. For the control options, see
the 'Details' in the help of \code{\link[stats]{optim}} for the possible arguments.}
}
\value{
Vector containing the estimates for kappa and lambda in the beta regression.
}
\description{
Function to run the Expectation-Maximization algorithm for the beta regression.
}
