% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_em_dbb.R
\name{EMrun_bet_dbb}
\alias{EMrun_bet_dbb}
\title{EMrun_bet_dbb}
\usage{
EMrun_bet_dbb(
  lam,
  z,
  v,
  mu,
  epsilon,
  link.precision,
  optim_method = "L-BFGS-B"
)
}
\arguments{
\item{lam}{initial values for the coefficients in lambda related to the precision parameter.}

\item{z}{response vector with 0 < z_i < 1.}

\item{v}{matrix containing the covariates for the precision submodel. Each column is a different covariate.}

\item{mu}{mean parameter (vector having the same size of z).}

\item{epsilon}{tolerance to controll convergence criterion.}

\item{link.precision}{a string containing the link function the precision parameter.
The possible link functions for the precision parameter are "identity", "log", "sqrt".}

\item{optim_method}{main optimization algorithm to be used. The available methods are the same as those of \code{optim} function. The default is set to "L-BFGS-B".}
}
\value{
Vector containing the estimates for lam in the beta regression.
}
\description{
Function (adapted for the discrimination test between bessel and beta - DBB) to run the Expectation-Maximization algorithm for the beta regression.
}
