% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_Bes.R
\name{Ew2z}
\alias{Ew2z}
\title{Ew2z}
\usage{
Ew2z(z, mu, phi)
}
\arguments{
\item{z}{response vector with 0 < z_i < 1.}

\item{mu}{mean parameter (vector having the same size of z).}

\item{phi}{precision parameter (vector having the same size of z).}
}
\value{
vector of expected values.
}
\description{
Auxiliary function required in the calculation of the
Fisher information matrix. It represents the conditional expected value E(W_i^s|Z_i), with s = -2; i.e.,
latent W_i^(-2) given the observed Z_i.
}
