% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_functions.R
\name{g_inv}
\alias{g_inv}
\title{g_inv}
\usage{
g_inv(sigma2, model)
}
\arguments{
\item{sigma2}{dispersion parameter.}

\item{model}{"bessel" or "beta"}
}
\description{
Function to obtain an approximation of the inverse of the function g() that relates the precision parameter to the dispersion parameter for bessel and beta distributions.
}
