% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_auxiliary_functions.R
\name{g_phi}
\alias{g_phi}
\title{g_phi}
\usage{
g_phi(phi, model)
}
\arguments{
\item{phi}{precision parameter.}

\item{model}{"bessel" or "beta"}
}
\description{
Function to obtain the function g() that relates the precision parameter to the dispersion parameter for bessel and beta distributions.
}
