% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_em_beta.R
\name{pred_accuracy_bet}
\alias{pred_accuracy_bet}
\title{pred_accuracy_bet}
\usage{
pred_accuracy_bet(
  residual,
  kap,
  lam,
  z,
  x,
  v,
  ntest,
  predict,
  link.mean,
  link.precision,
  em_controls,
  optim_method,
  optim_controls
)
}
\arguments{
\item{residual}{Character indicating the type of residual ("pearson", "score" or "quantile").}

\item{kap}{coefficients in kappa related to the mean parameter.}

\item{lam}{coefficients in lambda related to the precision parameter.}

\item{z}{response vector with 0 < z_i < 1.}

\item{x}{matrix containing the covariates for the mean submodel. Each column is a different covariate.}

\item{v}{matrix containing the covariates for the precision submodel. Each column is a different covariate.}

\item{ntest}{number of observations in the test set for prediction.}

\item{predict}{number of partitions (training and test sets) to be evaluated.}

\item{link.mean}{a string containing the link function for the mean.
The possible link functions for the mean are "logit","probit", "cauchit", "cloglog".}

\item{link.precision}{a string containing the link function the precision parameter.
The possible link functions for the precision parameter are "identity", "log", "sqrt".}

\item{em_controls}{a list containing two elements: \code{maxit} that contains the maximum number of iterations of the EM algorithm;
\code{em_tol} that defines the tolerance value to control the convergence criterion in the EM-algorithm.}

\item{optim_method}{main optimization algorithm to be used. The available methods are the same as those of \code{optim} function.}

\item{optim_controls}{a list of control arguments to be passed to the \code{optim} function in the optimization of the model. For the control options, see}
}
\value{
Vector containing the RSS for each partition of the full data set.
}
\description{
Function to calculate the Residual Sum of Squares for partitions (training and test sets) of
the data set. Residuals are calculated here based on the beta regression.
}
\seealso{
\code{\link{score_residual_bet}}, \code{\link{quantile_residual_bet}}, \code{\link{envelope_bet}}
}
