% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_functions.R
\name{print.bbreg}
\alias{print.bbreg}
\title{print.bbreg}
\usage{
\method{print}{bbreg}(x, ...)
}
\arguments{
\item{x}{object of class "bbreg" containing results from the fitted model.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Function providing a brief description of results related to the regression model (bessel or beta).
}
\examples{
\donttest{
fit = bbreg(agreement ~ priming + eliciting, data = WT)
fit}
}
\seealso{
\code{\link{fitted.bbreg}}, \code{\link{summary.bbreg}}, \code{\link{coef.bbreg}}, \code{\link{vcov.bbreg}}, \code{\link{plot.bbreg}}, \code{\link{predict.bbreg}}
}
