% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_DBB.R
\name{Qf_bes_dbb}
\alias{Qf_bes_dbb}
\title{Qf_bes_dbb}
\usage{
Qf_bes_dbb(lam, wz, z, v, mu)
}
\arguments{
\item{lam}{coefficients in lambda related to the covariates in v.}

\item{wz}{parameter wz representing E(1/W[i]|Z[i] = z[i], theta).}

\item{z}{response vector with 0 < z[i] < 1.}

\item{v}{matrix containing the covariates for the precision submodel. Each column is a different covariate.}

\item{mu}{mean parameter (vector having the same size of z).}
}
\value{
Scalar representing the output of this auxiliary function for the bessel case.
}
\description{
Auxiliary function (adapted for the discrimination test between bessel and beta - DBB) required in the Expectation-Maximization algorithm (Q-function related to the bessel model).
}
