% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Info_datasets.R
\docType{data}
\name{SA}
\alias{SA}
\title{Stress/Axiety data set}
\format{
Data frame containing 166 observations on 2 variables.
\describe{
\item{stress}{score, linearly transformed to the open unit interval.}
\item{anxiety}{score, linearly transformed to the open unit interval.}
}
}
\source{
Data can be obtained from the supplementary materials of \emph{Smithson and Verkuilen (2006)}. See also \emph{Barreto-Souza, Mayrink and Simas (2020)} for details.
}
\usage{
data(SA)
}
\description{
Stress and anxiety scores among nonclinical women in Townsville - Queensland, Australia.
}
\examples{
data(SA)
}
\references{
arXiv:2003.05157 (\href{https://arxiv.org/abs/2003.05157}{Barreto-Souza, Mayrink and Simas; 2020})

DOI:10.1037/1082-989X.11.1.54 (\href{https://psycnet.apa.org/record/2006-03820-004?doi=1}{Smithson and Verkuilen (2006)})
}
\keyword{datasets}
