% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_functions.R
\name{summary.bbreg}
\alias{summary.bbreg}
\title{summary.bbreg}
\usage{
\method{summary}{bbreg}(object, ...)
}
\arguments{
\item{object}{an object of class "bbreg" containing results from the fitted model.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Function providing a summary of results related to the regression model (bessel or beta).
}
\examples{
\donttest{
fit = bbreg(agreement ~ priming + eliciting|priming, data = WT)
summary(fit)}
}
\seealso{
\code{\link{fitted.bbreg}}, \code{\link{plot.bbreg}}, \code{\link{predict.bbreg}}
}
