% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Info_datasets.R
\docType{data}
\name{WT}
\alias{WT}
\title{Weather Task data set}
\format{
Data frame containing 345 observations on 3 variables.
\describe{
\item{agreement}{probability or the average between two probabilities indicated by each individual.}
\item{priming}{categorical covariate (0 = two-fold, 1 = seven-fold).}
\item{eliciting}{categorical covariate (0 = precise, 1 = imprecise).}
}
}
\source{
Data can be obtained from supplementary materials of \emph{Smithson et al. (2011)}. See also \emph{Barreto-Souza, Mayrink and Simas (2020)} for details.
}
\usage{
data(WT)
}
\description{
Weather task data set.
}
\examples{
data(WT)
}
\references{
arXiv:2003.05157 (\href{https://arxiv.org/abs/2003.05157}{Barreto-Souza, Mayrink and Simas; 2020})

DOI:10.1080/15598608.2009.10411918 (\href{https://www.tandfonline.com/doi/abs/10.1080/15598608.2009.10411918?tab=permissions&scroll=top}{Smithson and Verkuilen; 2009})

DOI:10.3102/1076998610396893 (\href{https://journals.sagepub.com/doi/abs/10.3102/1076998610396893?journalCode=jebb}{Smithson et al.; 2011})
}
\keyword{datasets}
