% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_em_bessel.R
\name{bessel}
\alias{bessel}
\alias{dbessel}
\alias{rbessel}
\alias{pbessel}
\alias{qbessel}
\title{Bessel Distribution}
\usage{
dbessel(z, mu = 1/2, phi = 1)

rbessel(n, mu = 1/2, phi = 1)

pbessel(q, mu = 1/2, phi = 1)

qbessel(p, mu = 1/2, phi = 1)
}
\arguments{
\item{z}{vector of numbers in (0,1) for which the p.d.f. is to be evaluated.}

\item{mu}{vector of numbers in (0,1) representing the mean parameter.}

\item{phi}{vector of positive numbers representing the precision parameter.}

\item{n}{sample size.}

\item{q}{vector of quantiles.}

\item{p}{vector of probabilities.}
}
\value{
scalar expressing the value of the density at z.
}
\description{
Functions to calculate the cumulative distribution, probability density,
generate bessel random numbers and find quantiles of the bessel distribution
}
\examples{
rbessel(100, mu = 0.2, phi = 10)
pbessel(0.4, mu = 0.1)
qbessel(0.8, mu = 0.8)
plot(dbessel)
}
\seealso{
\code{\link{simdata_bes}}, \code{\link{dbbtest}}, \code{\link{simdata_bet}}
}
