% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_DBB.R
\name{dbbtest.fit}
\alias{dbbtest.fit}
\title{dbbtest.fit}
\usage{
dbbtest.fit(z, x, v = NULL, epsilon = 10^(-5), link.mean, link.precision)
}
\arguments{
\item{z}{vector of response variables with length \code{n}. Each coordinate must belong to the standard unit interval (0,1).}

\item{x}{matrix of covariates with respect to the mean with dimension \code{(n,nkap)}.}

\item{v}{matrix of covariates with respect to the precision parameter. The default is \code{NULL}. If not \code{NULL} must be of dimension \code{(n,nlam)}.}

\item{epsilon}{tolerance value to control the convergence criterion in the Expectation-Maximization algorithm (default = 10^(-5)).}

\item{link.mean}{a string containing the link function for the mean.
The possible link functions for the mean are "logit","probit", "cauchit", "cloglog".}

\item{link.precision}{a string containing the link function the precision parameter.
The possible link functions for the precision parameter are "identity", "log", "sqrt".}
}
\value{
Object of class dbbtest, which is a list containing two elements. The 1st one is a table of terms
considered in the decision rule of the test; they are sum(z2/n) = sum_{i=1}^{n}(z_i^2)/n, sum(quasi_mu) = sum_{i=1}^{n}(tilde{mu_i}^2 + tilde{mu_i}(1-tilde{mu_i})/2)
|D_bessel| and |D_beta| as indicated in the main reference. The 2nd term of the list is the name of the selected model (bessel or beta).
}
\description{
Function to run the discrimination test between beta and bessel regressions (DBB).
}
\examples{
# Illustration using the Weather task data set available in the bbreg package.
n <- 100
x <- cbind(rbinom(n, 1, 0.5), runif(n, -1, 1))
v <- runif(n, -1, 1)
z <- simdata_bes(
  kap = c(1, -1, 0.5), lam = c(0.5, -0.5), x, v,
  repetition = 1, link.mean = "logit", link.precision = "log"
)
z <- unlist(z)
dbbtest.fit(z = z, x = x ,v = v, link.mean = "logit", link.precision = "identity")
}
\seealso{
\code{\link{simdata_bes}}, \code{\link{dbbtest}}, \code{\link{simdata_bet}}
}
