% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_DBB.R
\name{EMrun_bes_dbb}
\alias{EMrun_bes_dbb}
\title{EMrun_bes_dbb}
\usage{
EMrun_bes_dbb(lam, z, v, mu, epsilon, link.precision)
}
\arguments{
\item{lam}{initial values for the coefficients in lambda related to the precision parameter.}

\item{z}{response vector with 0 < z_i < 1.}

\item{v}{matrix containing the covariates for the precision submodel. Each column is a different covariate.}

\item{mu}{mean parameter (vector having the same size of z).}

\item{epsilon}{tolerance to controll convergence criterion.}

\item{link.precision}{a string containing the link function the precision parameter.
The possible link functions for the precision parameter are "identity", "log", "sqrt", "inverse".}
}
\value{
Vector containing the estimates for lam in the bessel regression.
}
\description{
Function (adapted for the discrimination test between bessel and beta - DBB) to run the Expectation-Maximization algorithm for the bessel regression.
}
