% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_Bet.R
\name{EMrun_bet}
\alias{EMrun_bet}
\title{EMrun_bet}
\usage{
EMrun_bet(kap, lam, z, x, v, epsilon)
}
\arguments{
\item{kap}{initial values for the coefficients in kappa related to the mean parameter.}

\item{lam}{initial values for the coefficients in lambda related to the precision parameter.}

\item{z}{response vector with 0 < z[i] < 1.}

\item{x}{matrix containing the covariates for the mean submodel. Each column is a different covariate.}

\item{v}{matrix containing the covariates for the precision submodel. Each column is a different covariate.}

\item{epsilon}{tolerance to control convergence criterion.}
}
\value{
Vector containing the estimates for kappa and lambda.
}
\description{
Function to run the Expectation-Maximization algorithm for the beta regression.
}
