% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_Bet.R
\name{EMrun_bet}
\alias{EMrun_bet}
\title{EMrun_bet}
\usage{
EMrun_bet(kap, lam, z, x, v, epsilon, link.mean, link.precision)
}
\arguments{
\item{kap}{initial values for the coefficients in kappa related to the mean parameter.}

\item{lam}{initial values for the coefficients in lambda related to the precision parameter.}

\item{z}{response vector with 0 < z_i < 1.}

\item{x}{matrix containing the covariates for the mean submodel. Each column is a different covariate.}

\item{v}{matrix containing the covariates for the precision submodel. Each column is a different covariate.}

\item{epsilon}{tolerance to control the convergence criterion.}

\item{link.mean}{a string containing the link function for the mean.
The possible link functions for the mean are "logit","probit", "cauchit", "cloglog".}

\item{link.precision}{a string containing the link function the precision parameter.
The possible link functions for the precision parameter are "identity", "log", "sqrt", "inverse".}
}
\value{
Vector containing the estimates for kappa and lambda in the beta regression.
}
\description{
Function to run the Expectation-Maximization algorithm for the beta regression.
}
