% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_Bes.R
\name{Ew1z}
\alias{Ew1z}
\title{Ew1z}
\usage{
Ew1z(z, mu, phi)
}
\arguments{
\item{z}{response vector with 0 < z[i] < 1.}

\item{mu}{mean parameter (vector having the same size of z).}

\item{phi}{precision parameter (vector having the same size of z).}
}
\value{
Vector of expected values.
}
\description{
Auxiliary function required in the Expectation-Maximization algorithm (E-step) and in the calculation of the
Fisher information matrix. It represents the conditional expected value E(W[i]^s|Z[i]), with s = -1; i.e.,
latent W[i]^(-1) given the observed Z[i].
}
