% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_DBB.R
\name{gradlam_bes_dbb}
\alias{gradlam_bes_dbb}
\title{gradlam_bes_dbb}
\usage{
gradlam_bes_dbb(lam, wz, z, v, mu, link.precision)
}
\arguments{
\item{lam}{coefficients in lambda related to the covariates in v.}

\item{wz}{parameter wz representing E(1/W_i|Z_i = z_i, theta).}

\item{z}{response vector with 0 < z_i < 1.}

\item{v}{matrix containing the covariates for the precision submodel. Each column is a different covariate.}

\item{mu}{mean parameter (vector having the same size of z).}

\item{link.precision}{a string containing the link function the precision parameter.
The possible link functions for the precision parameter are "identity", "log", "sqrt", "inverse".}
}
\value{
Scalar representing the output of this auxiliary gradient function for the bessel case.
}
\description{
Gradient of the Q-function (adapted for the discrimination test between bessel and beta - DBB) to calculate the gradient required for optimization via \code{optim}.
This option is related to the bessel regression.
}
