% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_Bet.R
\name{infmat_bet}
\alias{infmat_bet}
\title{infmat_bet}
\usage{
infmat_bet(theta, z, x, v)
}
\arguments{
\item{theta}{vector of parameters (all coefficients: kappa and lambda).}

\item{z}{response vector with 0 < z[i] < 1.}

\item{x}{matrix containing the covariates for the mean submodel. Each column is a different covariate.}

\item{v}{matrix containing the covariates for the precision submodel. Each column is a different covariate.}
}
\value{
Vector of standard errors.
}
\description{
Function to compute standard errors based on the Fisher information matrix for the beta regression.
}
