% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_Bes.R
\name{Qf_bes}
\alias{Qf_bes}
\title{Qf_bes}
\usage{
Qf_bes(theta, wz, z, x, v)
}
\arguments{
\item{theta}{vector of parameters (all coefficients: kappa and lambda).}

\item{wz}{parameter representing E(1/W[i]|Z[i] = z[i], theta).}

\item{z}{response vector with 0 < z[i] < 1.}

\item{x}{matrix containing the covariates for the mean submodel. Each column is a different covariate.}

\item{v}{matrix containing the covariates for the precision submodel. Each column is a different covariate.}
}
\value{
Scalar representing the output of this auxiliary function for the bessel case.
}
\description{
Auxiliary function required in the Expectation-Maximization algorithm (Q-function related to the bessel model).
}
