% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_DBB.R
\name{Qf_bet_dbb}
\alias{Qf_bet_dbb}
\title{Qf_bet_dbb}
\usage{
Qf_bet_dbb(lam, phiold, z, v, mu)
}
\arguments{
\item{lam}{coefficients in lambda related to the covariates in v.}

\item{phiold}{previous value of the precision parameter (phi).}

\item{z}{response vector with 0 < z[i] < 1.}

\item{v}{matrix containing the covariates for the precision submodel. Each column is a different covariate.}

\item{mu}{mean parameter (vector having the same size of z).}
}
\value{
Scalar representing the output of this auxiliary function for the beta case.
}
\description{
Auxiliary function (adapted for the discrimination test between bessel and beta - DBB) required in the Expectation-Maximization algorithm (Q-function related to the beta model).
}
