% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_Bet.R
\name{quantile_residual_bet}
\alias{quantile_residual_bet}
\title{quantile_residual_bet}
\usage{
quantile_residual_bet(kap, lam, z, x, v)
}
\arguments{
\item{kap}{coefficients in kappa related to the mean parameter.}

\item{lam}{coefficients in lambda related to the precision parameter.}

\item{z}{response vector with 0 < z[i] < 1.}

\item{x}{matrix containing the covariates for the mean submodel. Each column is a different covariate.}

\item{v}{matrix containing the covariates for the precision submodel. Each column is a different covariate.}
}
\value{
Vector containing the quantile residuals.
}
\description{
Function to calculate quantile residuals based on the beta regression. Details about this type of residual can be found in \emph{Pereira (2019)}.
}
\references{
DOI:10.1080/03610918.2017.1381740 (\href{https://www.tandfonline.com/doi/abs/10.1080/03610918.2017.1381740}{Pereira; 2019})
}
