% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_functions.R
\name{startvalues}
\alias{startvalues}
\title{startvalues}
\usage{
startvalues(z, x, v)
}
\arguments{
\item{z}{response vector with 0 < z[i] < 1.}

\item{x}{matrix containing the covariates for the mean submodel. Each column is a different covariate.}

\item{v}{matrix containing the covariates for the precision submodel. Each column is a different covariate.}
}
\description{
Function providing initial values for the Expectation-Maximization algorithm.
}
